+++ src/googleApi/GoogleCreateTask.ts
@@ -52,7 +52,7 @@ export async function CreateGoogleTask(
 export async function CreateGoogleTaskFromOldTask(
 	plugin: GoogleTasks,
 	newTask: Task
-) {
+): Promise<boolean> {
 	const requestHeaders: HeadersInit = new Headers();
 	requestHeaders.append(
 		"Authorization",
@@ -81,8 +81,11 @@ export async function CreateGoogleTaskFromOldTask(
 			await response.json();
 
 			plugin.refreshScheduleViews();
+			return true;
 		}
+		return false;
 	} catch (error) {
 		console.error(error);
+		return false;
 	}
 }